Review comments
Use the single call method that Webb suggested for the mock response.
I left out the BufferedReader case because I don't understand it.
I left in the Exception case although I don't understand it either,
but I'm guessing it can help to unify some of the "abnormal" test cases
with the "normal" one - but it is not used in this incarnation.