Avoid Sync message overflow
PBENCH-1120

A SQL error was observed in deployment where `pbench-index` logged an error on
the `INDEX` sync object because a tarball was somehow not present. The message
string generated by `indexing_tarballs.py` exceeded the `VARCHAR(255)` column
specification.

This isn't an attempt to address the root problem, but to address the symptom
of overloading the operation table message column in the future so at least
errors are properly recorded.

This reworks some of the `indexing_tarballs.py` messages to avoid redundancy
(e.g., naming the dataset or tarball isn't necessary as the records are linked
to the `Dataset`), but also removes the limit on the message column as a
precaution.