Add `AuthToken.API_KEY` SQL ENUM
PBENCH-1136

Stage one: this doesn't fix the API key encryption problem (Nikhil is working
on that), but this adds the missing SQL ENUM value.

Rather than creating a new Alembic revision file (which would be the normal
process) I decided to keep this simple by updating the general API KEY
revision file. As we have no database deployments we'll need to upgrade over
this period, this strategy avoids adding another revision to the tree.