Add API Key functional testing (#3401)
PBENCH-1135

Add mechanism and a test case to verify that we can authenticate and identify
the client user from an API key, simplistically using `GET /datasets?mine`.

This also adds a new `alembic-migration` command, `show` to display what
alembic sees as the "heads" and "history" of the revision chain. I added this
when I accidentally constructed multiple heads and got weird errors while I
was building the `AuthType.API_KEY` upgrade, but didn't include it in that
final PR. It seems useful, so I'm sweeping it in here.