Fix the image tag
Fixes #3445

pbench_demo uses an image downloaded from Quay.io. It assumes that the
image will contain `pbench-generate-token` which is available
currently but it is deprecated and will not be available in the
future.  In particular, the specified image
`quay.io/pbench/pbench-agent-all-fedora-36:all` does *not* contain the
script. We need to make use
`quay.io/pbench/pbench-agent-all-fedora-36:b0.72` instead.

PBENCH-1129