Add/fix tests
This commit adds a "functional" test for pbench-list-tools to verify
the output when labeled hosts have registered tools.

It also fixes an incorrect legacy test (test-07) for
pbench-postprocess-tools: the test was using a labeled host, but it
was testing the wrong tool output directory (without the label).

It adds a similar legacy test (test-07.1) to test the unlabeled host
case.