Fix dashboard logout (#3468)
PBENCH-1185

Starting relatively recently, Keycloak has a distinct callback URI setting for
"post logout", which I noticed while debugging my Postman configuration on
the staging server. Adding the proper "post logout callback URI" solved the
problem and allowed a clean logout. Then it only took me an hour of searching
to figure out how to specify this in the REST call; to say that it's "not
excessively well documented" is understanding the point ...