Work around requests version conflict
PBENCH-1190

The Pbench Agent depends on a version of the `requests` package of at least
2.27 for the `requests.exceptions.JSONDecodeError` exception. We'd like to
require 2.31 as that resolves a CVE, however Pbench Agent supports Python
3.6 which won't work with 2.31. Adding to this mix, the `pquisby` package
requires exactly 2.25.

For now, in order to move forward with Quisby, we remove the version
constraints from `requests` for Pbench and remove the reference to the
`JSONDecodeError` exception.

(Note that a secondary level of mess here is that the Pbench CI installs
*all* our `requirements*.txt` files in the same environment, even though
Pbench Agent and Pbench Server don't really operate in the same environment
and should be completely separate. This currently prevents us from having
incompatible requirements between the two and we ought to change that
going forward.