Add URI to TOC API
PBENCH-624

We're not quite ready to rebuild the `/datasets/{dataset}/contents` API on the
cache manager. (I actually tried; it's not all there, and currently would need
to be built via a full tarball unpack each time we build a `CacheManager`
instance, while the changes to persist it via SQL or Redis, which I'd like to
do soon, seemed too big a lift right now.)

However, we have planned for some time to supplement the TOC information with
a uri element for each file and subdirectory; this much smaller change makes the
code match the API documentation.