Harmonize "inventory" APIs
PBENCH-1194

Rework Tarball inventory access to avoid the nascent cache map so that we can
rely on this inside the Pbench Server process where tarballs haven't been
unpacked.

This also fixes some incompatibilities in the Quisby APIs which relied on the
older extract semantics.

In order to avoid stale streams on the `inventory` API I encapsulated the
`extractfile` stream into an object that can behave as a stream but will close
both the stream and tarball when done: this is deferred until Flask tears down
the Response object.

Finally, I've added functional tests validating the `contents`, `inventory`,
`visualize`, and `compare` APIs so we'll catch any further infrastructure
"drift" that might affect them.