Provide dataset summary info on upload
PBENCH-1204

This stems from a request from the UI to help optimize a partial refresh after
using the relay upload dialog, by providing identification of the new dataset.
Although a client using the traditional `PUT /upload` already knows the name
and MD5, the returned information may be helpful.

This is a DRAFT partly because I added a validation of the new information to
the functional test, which I've renamed in #3473 ... I'll do that merge here
after it's gone in.

It's also DRAFT because while I like the idea of including URIs (and in
particular this addresses a certain request regarding accessibility of the
tarball), I'm not really sure which URIs to include or in what form. I'm
certain that one or two people might possibly have opinions on this subject!