Fix server RPM builds on Fedora (b0.69)
python3.9 is needed as a build requirement. That may be required
in b0.72 and main as well, but we'll cross that bridge when/if we need to.

pre-commit was failing with weird errors. After clearing the cache, I
was asked to login to gitlab: the config was different from main and
was trying to get flake8 from gitlab. I copied the config from main
(which is using github for both flake8 and black) into the b0.69
branch and that seems to work.