Identify benchmark type (#3495)
PBENCH-1210

Right now we only support visualizing and comparing `uperf` benchmark data
with Quisby. However as we expand that support within the server, the UI code
will need to be able to identify the schema of the output data, which depends
on the benchmark.

As a simple solution, add a `"benchmark"` field to the response data which the
client can read.