Improve client benchmark identification (#3496)
PBENCH-1186
PBENCH-1187

A bit of a catch-all. First, this simplifies and standardizes the dashboard's
need to identify a benchmark. Eventually we'll be "smart enough" to go beyond
the Pbench benchmark wrapper (as we intend to move away from relying on them)
and `dataset.metalog.pbench.script` will become unreliable. This adds a new
`server.benchmark`, which right now is the same but can be extended beyond
that in the future.

The dashboard would also like to be able to identify which benchmark types the
current server can support, in advance, and we'd like to avoid hardcoding that
knowledge in the dashboard. For this purpose, I've added a "visualization" key
in the endpoints response to indicate the benchmark types supported by the
current server's `datasets_visualize` and `datasets_compare` APIs.

For some reason, minor consolidation of metadata setting code in the intake
API base resulted in weird unit test and functional test failures. The odd
part is that these should have been failing earlier, as they're cases where
a dataset doesn't have `server.index-map` but the API requires it (for example
in `datasets_detail` and `datasets_contents`). I twizzled the code a bit to
fail with a `CONFLICT` error (which had already been used elsewhere) rather
than a server internal error when the dataset is `server.archiveonly`, and
adjusted the tests to match.