Fix some problems with archive-only behavior  (#3497)
* Fix some problems with archive-only behavior

PBENCH-1215

Webb pointed out a broken functional test in #3496. Once it was testing what I
had wanted it to test, I realized that it wasn't quite working. This contains
some server and functional test changes to make archive-only dataset queries
work correctly and to be sure that's adequately tested in the functional test
cases.