"Note" non-fatal INTAKE behaviors
PBENCH-1231

After accidentally observing an uploaded dataset that was marked "archiveonly"
not indexed (because the file name didn't match the result directory name), we
decided that the intake process should make the reasons more obvious.

I've handled this by adding "notes", which are recorded both in the successful
JSON response payload and in the finalized audit record. While I was at it, I
exposed our decisions regarding the canonical benchmarked workload name and
the computed expiration date.

While doing this, I stripped the rather prolific logging of `INTAKE` to just
two, a "pre" log identifying the file and user, and a "post" log providing the
summary file system status (adding the size of the tarball).

I debated logging the "notes", but didn't. On one hand, it would make the info
more easily visible; on another it restores verbosity, and the notes strings
are going to be unwieldy in the log. I've also several times considered that
we might log every audit record, which would include attributes... but might
be hard to read in the linear log format.

In any case, it's easy enough to do an Audit query on the dataset resource_id
even when the client doesn't capture the "notes".