"Note" non-fatal INTAKE behaviors (#3510)
* "Note" non-fatal INTAKE behaviors

PBENCH-1231

After accidentally observing an uploaded dataset that was marked "archiveonly"
not indexed (because the file name didn't match the result directory name), we
decided that the intake process should make the reasons more obvious.

I've handled this by adding "notes", which are recorded both in the successful
JSON response payload and in the finalized audit record. While I was at it, I
exposed our decisions regarding the canonical benchmarked workload name and
the computed expiration date.

While doing this, I stripped the rather prolific logging of `INTAKE` to just
two, a "pre" log identifying the file and user, and a "post" log providing the
summary file system status (adding the size of the tarball).