Allow deletion of uploaded relay datasets
PBENCH-1238

As a stopgap for "final" Relay support under the agent, we plan to establish a
known project Relay Server (e.g., on our AWS host) for testing. That server
has limited storage, and substantial use. To avoid requiring constant manual
management of space, our plan is to implement a mechanism to cause the Pbench
Server to delete files from the Relay Server after intake, and to change the
Pbench Dashboard relay action to always supply that option. Note that this PR
adds the `?delete=t` to the dashboard *unconditionally*; in the future we can
consider adding a checkbox to allow users to control the behavior, especially
when they own a private relay server.