Move index-map out of metadata
PBENCH-462

This moves the Elasticsearch document index map out of the `server.index-map`
metadata key and gives it its own SQL table. There are more changes than I'd
expected to get this all back together. I'm not entirely sure I'm "done", but
I want to get this up.

For one thing, I'd considered refactoring the bulk operation to drive a
generator out of the SQLAlchemy `.all()` query instead of building, returning,
and then traversing a two-level `dict`. I'd also like to try to *upgrade* an
existing dataset by translating `server.index-map` into `IndexMap` rows,
ideally in the Alembic upgrade: although it probably makes more sense to just
throw away (or re-import) existing datasets on the staging server.

And, just to get the CI to run, this locks `pquisby<0.0.13` to avoid the
import bug; but I'm hoping that will be fixed before I submit, which is the
main reason to post this as a draft.