Second commit.
This removes the `map` method, replacing it with an `exists` boolean and a
`stream` to generate index/id pairs for the bulk generators. It also
simplifies the `find` to an `indices` method returning just a list of
matching full indices. These substantially simplify the calling code points.

Added full unit testing of the index_map module. Note that while I had the
functional tests working, I'm pushing this commit a bit blindly as my
`runlocal` is having trouble with the Red Hat CDN this weekend; maybe the
CI will work from the lab.

I also added a prototype `move-index.py` program which (with a few simple
changes) can be used to replace all the `server.index-map` metadata on the
staging server (or production server) with `IndexMap` rows. Right now, it's
in a simple form that can be run on the staging server to *report* all the
index data, which helped in debugging. Once I add the `IndexMap` model
imports it'll no longer be able to run on staging until we deploy, whereas
it's hard to test the basic discovery on a `runlocal` as there's no longer
any `server.index-map` metadata. I could have named it `catch-22.py`...