Adapt to Quisby fio support
PBENCH-1212

(This Jira issue was originally a placeholder in case changes had to be made
to the `comparison` or `visualize` APIs for `fio` support. It seemed likely
that wouldn't be necessary, and I'm appropriating the Jira to unlock the
`pquisby` package version and adjust tests.)

First, unlike the `pquisby` package version to allow consuming the 0.0.17
version with `fio` support added.

Both unit and functional tests validate the set of benchmarked workloads for
which Pbench claims support, and needed to be updated for the addition of
`fio`. The `endpoints` API was also reporting the workloads in order from
Pquisby's ENUM, and to make the tests a bit more supportable I sorted the
list we report.