Move index-map out of metadata (#3519)
* Move index-map out of metadata

PBENCH-462

This moves the Elasticsearch document index map out of the `server.index-map`
metadata key and gives it its own SQL table. There are more changes than I'd
expected to get this all back together. This refactors the bulk operation to drive a
`stream` generator out of the SQLAlchemy `.all()` query instead of building,
returning, and then traversing a two-level `dict`, and adds a quick `exists` check
and `indices` to return a simple list of index names, again to avoid needing to
process a `dict`.

There's also a utility script to convert existing `server.index-map` metadata into
`IndexMap` DB rows, which can be run one time on an upgrade. We'll probably
do this exactly twice: once for the staging server and then for the production
server.