Another patch ...
I ran the fixed tool again on the staging server to delete the `server.index-map`
metadata (which is under a separate `--delete` option out of paranoia), and then
again to verify it was "fixed" and got an error... because the code didn't
consider that there might not *be* `server.index-map` metadata values. Might
as well fix that, too.