Fix the SQL constraint violation logic (#3528)
* Fix the SQL constraint violation logic

PBENCH-1251

This is mostly a fix for the ops review problem discovered Thursday, that our
carefully unit-tested logic to decode SQLAlchemy `IntegrityError` instances to
identify the specific violated constraint (specifically, duplicate vs null)
doesn't work in a production environment.

The root problem is that the text strings reported differ between the unit
test environment (`sqlite3` engine) and the production environment
(`postgresql` engine).

In order to consolidate on the central constraint error decoder, (which was
why this wasn't done earlier), we need all SQL-related exception constructors
to be consistent, but I didn't like dropping all context-specific arguments.
Instead I've adopted a `**kwargs` model where the decoder passes through
arbitrary parameters, which are captured in the exception to provide context.

This also adopts a more "pythonic exception" model where `e.args` captures a
statically formatted message string and allows the base `Exception.__str__` to
provide formatting. I've done this for all the SQLAlchemy model objects, but
nowhere else.

And to avoid complication in many places where we use `str(exception)`,
I'm not including the `kwargs` or raw `cause` in the `args` tuple. I played
with that idea, but since `Exception.__str__` formats the entire tuple, the
results are ugly and also more difficult to meaningfully compare in tests. We
use `str(exception)` a lot both internally and in tests.